+++ packages/terra-framework-docs/CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * Added
   * Added documentation for hyperlinks in `terra-collapsible-menu-view`.
-  
+  * Added doc and test example for `collapsible-menu-view` multi-select group items.
+
 * Updated
   * Updated examples for `terra-collapsible-menu-view` to use more meaningful controls and control labels.